Style: style SCSS
