+++ src/css/detailImageSlide.scss
@@ -1,14 +1,15 @@
 .detailImageContainerPerPlaceWrap{
   display: flex;
   align-items: center;
+  .detail_categoryslide-imagecontent{
+    width:343px;
+    height:256px;
+    position: relative;
+    img {  
+      object-fit: cover;
+      width: 100%;
+      height: 100%;
+    }
+  }
 }
-.detail_categoryslide-imagecontent{
-  width:343px;
-  height:256px;
-  position: relative;
-}
-.detail_categoryslide-imagecontent > img {  
-  object-fit: cover;
-  width: 100%;
-  height: 100%;
-}
+
