+++ src/css/imageSlide.scss
@@ -2,17 +2,20 @@
   display: flex;
   align-items: center;
   z-index: 0;
+
+  .write_categoryslide-imagecontent{
+    width:343px;
+    height:256px;
+    position: relative;
+
+    img {  
+      object-fit: cover;
+      width: 100%;
+      height: 100%;
+    }
+  }
 }
-.write_categoryslide-imagecontent{
-  width:343px;
-  height:256px;
-  position: relative;
-}
-.write_categoryslide-imagecontent > img {  
-  object-fit: cover;
-  width: 100%;
-  height: 100%;
-}
+
 .addButton{
   width: 343px;
   height: 40px;
@@ -25,4 +28,8 @@
   align-items: center;
   justify-content: center;
   margin-top:8px;
+
+  input{
+    display: none;
+  }
 }
