+++ src/css/kakaomap.scss
@@ -1,0 +1,5 @@
+/* 지도 */
+.writeMapWrap{
+  width: 100vw;
+  height: 370px;
+}
