+++ src/css/regionModal.scss
@@ -1,25 +1,8 @@
-.background{
-  position: fixed;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  /* background-color: rgba(0,0,0,0.50); */
-  z-index: 0;
-}
-.region_wrap{
-  position: fixed;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  max-height: 80%;
-  width: 330px;
-  height: 321px;
-  padding: 18px 24px;
-  /* box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1); */
-  background:#fff;
-  border-radius: 10px;
-  text-align: center;
+.regionWrap.openModal {
+  display: flex;
+  align-items: center;
+  /* 팝업이 열릴때 스르륵 열리는 효과 */
+  animation: modal-bg-show 0.3s;
 }
 .regionWrap{
   display: none;
@@ -30,76 +13,92 @@
   left: 0;
   z-index: 99;
   background-color: rgba(0, 0, 0, 0.6);
-}
-.regionWrap>section {
-  width: 90%;
-  max-width: 560px;
-  margin: 0 auto;
-  border-radius: 12px;
-  background-color: #fff;
-  /* 팝업이 열릴때 스르륵 열리는 효과 */
-  animation: modal-show 0.3s;
-  overflow: hidden;
-}
-.modalTitle{
-  font-weight: 700;
-  font-size: 16px;
-  line-height: 23px;
-  text-align: left;
-  margin-bottom: 18px;
-  color:#606773;
-}
-.regionWrap.openModal {
-  display: flex;
-  align-items: center;
-  /* 팝업이 열릴때 스르륵 열리는 효과 */
-  animation: modal-bg-show 0.3s;
-}
-.regionsWrap{
-  display: flex;
-  flex-wrap: wrap;
-  grid-column-gap: 18px;
-  grid-row-gap: 14px;
-  width: 282px;
-}
-._regions{
-  width: 82px;
-  height: 28px;
-  border-radius: 10px;
-  display:flex;
-  flex-wrap: nowrap;
-  justify-content: center;
-  align-items: center;
-  background: #F5F9FF;
-  box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.1);
-  border-radius: 10px;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 20px;
-}
 
-.regions input{
-  display: none;
-}
+  .background{
+    position: fixed;
+    top: 0;
+    left: 0;
+    bottom: 0;
+    right: 0;
+    z-index: 0;
 
+    .region_wrap{
+      position: fixed;
+      left: 50%;
+      top: 50%;
+      transform: translate(-50%, -50%);
+      max-height: 80%;
+      width: 330px;
+      height: 321px;
+      padding: 18px 24px;
+      background:#fff;
+      border-radius: 10px;
+      text-align: center;
 
+      section {
+        max-width: 560px;
+        margin: 0 auto;
+        border-radius: 12px;
+        background-color: #fff;
 
-.doneButton{
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  margin-top: 16px;
-}
-.close{
-  width: 258px;
-  height: 36px;
-  background: #8ACEFF;
-  border-radius: 8px;
-  color: #F5F9FF;
-  font-weight: 700;
-  font-size: 14px;
-  line-height: 20px;
-  border: none;
+        .modalTitle{
+          font-weight: 700;
+          font-size: 16px;
+          line-height: 23px;
+          text-align: left;
+          margin-bottom: 18px;
+          color:#606773;
+        }
+
+        .regionsWrap{
+          display: flex;
+          flex-wrap: wrap;
+          grid-column-gap: 18px;
+          grid-row-gap: 14px;
+          width: 282px;
+
+          ._regions{
+            width: 82px;
+            height: 28px;
+            border-radius: 10px;
+            display:flex;
+            flex-wrap: nowrap;
+            justify-content: center;
+            align-items: center;
+            background: #F5F9FF;
+            box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.1);
+            border-radius: 10px;
+            font-weight: 400;
+            font-size: 14px;
+            line-height: 20px;
+          
+            input{
+              display: none;
+            }
+          }
+        }
+
+        .doneButton{
+          width: 100%;
+          display: flex;
+          justify-content: center;
+          margin-top: 16px;
+          .close{
+            width: 258px;
+            height: 36px;
+            background: #8ACEFF;
+            border-radius: 8px;
+            color: #F5F9FF;
+            font-weight: 700;
+            font-size: 14px;
+            line-height: 20px;
+            border: none;
+          }
+        }
+      }
+    }
+  }
+
 }
 
 @keyframes modal-show {
